llama-index/hello-hf/requirements.txt
llama-index-llms-huggingface-api
llama-index-embeddings-huggingface
python-dotenv